backend/requirements.txt
flask
flask_cors
flask_socketio
folium
polyline
python-dotenv